Moved night_writer.rb contents into translator.rb, added comments to night_writer.rb regarding file transfer information from video and cleared out the message.txt file.
