--- lib/message.txt
@@ -1,1 +1,0 @@
-Message for demonstration purposes.
